rename autorization page to review
